refactor procedure.py
